Fixing a resync bug.
Before synchronizing the data between the plugin and EOS, the plugin sends the
'sync start'. As part of this command, EOS returns a timestamp value and the
plugin sets an internal variable to this value. In the next step of the sync,
the plugin checks whether the last updated value is that same as that returned
by EOS. Since they will always be the same, the sync is completely bypassed.
The patch passes a flag to _run_openstack_cmd() which tells the method to
ignore the returned value.